# Precompute grids and coeffs