@oxinabox 2020-02-02
From stacking this ontop of #60:
```
julia> @btime grad($_fdm, $f, $xs);
  193.883 ms (51768 allocations: 1.93 MiB)

julia> @btime grad($_fdm5, $f, $xs);
  400.359 ms (51768 allocations: 1.84 MiB)
```

With #60 alone:
```
julia> @btime grad($_fdm, $f, $xs);
  200.205 ms (49777 allocations: 2.66 MiB)

julia> @btime grad($_fdm5, $f, $xs);
  398.832 ms (59732 allocations: 3.87 MiB)
```

interesting that it seems to make the `central_fdm(2)` case slightly worse.
Not sure whats up with that.