@oxinabox 2020-03-28
> Instead of precomputing you could memoize. Avoids having to determine upfront what exact combinations should be precomputed. Taking the cost once per session is probably ok.

Yeah, good idea.

I think I might have looked into it, and struggled to workout how to do it without getting allocations.
The precomputes use tuples which are very compiler friendly.

But I can't recall if I did, so I should check again